requirements.txt
mysql-connector
mysql-connector-python
customtkinter
bcrypt
matplotlib
CTkMessagebox
Pillow